# modules/portfolio/saas/api-contract.yaml
openapi: 3.0.3
info:
  title: Portfolio Module API
  description: Investment governance engine for venture portfolio analysis, allocation, and decision-making
  version: 1.0.0
  contact:
    name: BruceAI Platform Team
    email: platform@bruceai.com
servers:
  - url: https://api.bruceai.com/v1
    description: Production API
security:
  - bearerAuth: []
paths:
...
  /portfolio/ventures:
    get:
      summary: Get All Portfolio Ventures
      description: List all ventures in portfolio with current health metrics
      operationId: getPortfolioVentures
      tags:
        - Ventures
      parameters:
        - name: health_threshold
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
          description: Filter to ventures with health score below this threshold
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [health_score, funding, age]
            default: health_score
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of ventures
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Venture'
                  summary:
                    type: object
                    properties:
                      total_ventures:
                        type: integer
                      average_health_score:
                        type: number
                      ventures_at_risk:
                        type: integer
  /portfolio/ventures/{venture_id}:
    get:
      summary: Get Venture Details
      description: Retrieve detailed information for a specific venture
      operationId: getVenture
      tags:
        - Ventures
      parameters:
        - name: venture_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Venture details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentureDetail'
        '404':
          description: Venture not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token issued by BruceAI auth service
  schemas:
...
    Venture:
      type: object
      properties:
        venture_id:
          type: string
        name:
          type: string
        health_score:
          type: number
          minimum: 0
          maximum: 100
        funding_amount:
          type: number
        monthly_burn:
          type: number
        runway_months:
          type: number
          nullable: true
        market_traction:
          type: number
          minimum: 0
          maximum: 100
        team_strength:
          type: number
          minimum: 0
          maximum: 100
        last_updated:
          type: string
          format: date-time
    VentureDetail:
      allOf:
        - $ref: '#/components/schemas/Venture'
        - type: object
          properties:
            risk_factors:
              type: array
              items:
                type: string
            allocation_percent:
              type: number
              minimum: 0
              maximum: 100
            last_decision:
              $ref: '#/components/schemas/GovernanceDecision'
            metrics_trend:
              type: object
              additionalProperties:
                type: array
                items:
                  type: number
    GovernanceDecision:
      type: object
      properties:
        decision_id:
          type: string
        venture_id:
          type: string
        decision_type:
          type: string
          enum: [scale, iterate, pause, kill]
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
        rationale:
          type: string
        previous_decision:
          type: string
          enum: [scale, iterate, pause, kill]
          nullable: true
        agent_id:
          type: string
        recorded_at:
          type: string
          format: date-time
        recorded_by:
          type: string
          description: User or agent that recorded the decision
        review_id:
          type: string
...